v1.0.3 Update
1.	Updated CandiceAIManager
a.	Fixed CandiceTile bug
2.	Updated AI Controller
a.	Removed Callback parameter (_onAttackComplete) from DealTimedDamage and FireProjectile methods/functions.
i.	The callback that was registered in the constructor of the CandiceModuleCombat class is now the only one used.
b.	Removed onAttackComplete() function in AIController.
i.	onAttackComplete(bool success) is still available in the CandiceAIController.